Split IO tests as much as possible
Rest of io_test.py tests both loader and saver.
