+++ mantidimaging/core/io/test/utility_test.py
@@ -1,0 +1,45 @@
+from __future__ import absolute_import, division, print_function
+
+import logging
+import os
+
+from mantidimaging.helper import initialise_logging
+from mantidimaging.core.configs.recon_config import ReconstructionConfig
+from mantidimaging.core.io import utility
+from mantidimaging.test.file_outputting_test_case import (
+        FileOutputtingTestCase)
+
+
+class UtilityTest(FileOutputtingTestCase):
+    def __init__(self, *args, **kwargs):
+        super(UtilityTest, self).__init__(*args, **kwargs)
+
+        # force silent outputs
+        initialise_logging()
+        self.config = ReconstructionConfig.empty_init()
+        self.config.func.log_level = logging.CRITICAL
+
+    def test_get_candidate_file_extensions(self):
+        self.assertEquals(
+                ['tif', 'tiff'],
+                utility.get_candidate_file_extensions('tif'))
+
+        self.assertEquals(
+                ['tiff', 'tif'],
+                utility.get_candidate_file_extensions('tiff'))
+
+        self.assertEquals(
+                ['png'],
+                utility.get_candidate_file_extensions('png'))
+
+    def test_get_file_names(self):
+        # Create test file with .tiff extension
+        tiff_filename = os.path.join(self.output_directory, 'test.tiff')
+        with open(tiff_filename, 'wb') as tf:
+            tf.write(b'\0')
+
+        # Search for files with .tif extension
+        found_files = utility.get_file_names(self.output_directory, 'tif')
+
+        # Expect to find the .tiff file
+        self.assertEquals([tiff_filename], found_files)
